# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Inventario",
        "description": "Documentación oficial de la API para la gestión de inventarios, áreas y bienes.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
        "/areas": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Listar todas las áreas",
                "description": "Retorna un listado de áreas con sus departamentos, edificio y responsable.",
                "operationId": "b35f6f18c1157ebddfba113fe246bf90",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Area"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Areas"
                ],
                "summary": "Crear una nueva área",
                "operationId": "e3b3bd99f2014fd763018a030be4fd38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del área",
                                        "type": "string",
                                        "example": "Almacén B"
                                    },
                                    "description": {
                                        "description": "Descripción opcional",
                                        "type": "string",
                                        "example": "Sector de fríos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Área creada exitosamente"
                    }
                }
            }
        },
        "/areas/{id}": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Obtener detalles de un área específica",
                "description": "Retorna la información del área junto con su edificio, responsable y departamentos asociados.",
                "operationId": "89b34cddfab1b7c50aba560bfbe01069",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del área encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "area_nombre": {
                                            "type": "string",
                                            "example": "Dirección de TI"
                                        },
                                        "area_codigo": {
                                            "type": "string",
                                            "example": "TI-001"
                                        },
                                        "edificio": {
                                            "description": "Datos del edificio (cargado con load)",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Edificio Central"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "responsable": {
                                            "description": "Datos del responsable (cargado con load)",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "res_nombre": {
                                                    "type": "string",
                                                    "example": "Carlos"
                                                },
                                                "res_apellidos": {
                                                    "type": "string",
                                                    "example": "López"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "departamentos": {
                                            "description": "Lista de departamentos asociados",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 45
                                                    },
                                                    "dep_nombre": {
                                                        "type": "string",
                                                        "example": "Soporte Técnico"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Área no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Areas"
                ],
                "summary": "Actualizar un área existente",
                "operationId": "2a84eb6438724f670b6c0e3cce2544e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del área a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del área a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "area_nombre",
                                    "area_codigo"
                                ],
                                "properties": {
                                    "area_nombre": {
                                        "description": "Nombre del área",
                                        "type": "string",
                                        "example": "Dirección General"
                                    },
                                    "area_codigo": {
                                        "description": "Código único del área",
                                        "type": "string",
                                        "example": "DG-001"
                                    },
                                    "id_resguardante_responsable": {
                                        "description": "ID del usuario subdirector responsable",
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "id_edificio": {
                                        "description": "ID del edificio asociado",
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Área actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "area_nombre": {
                                            "type": "string",
                                            "example": "Dirección General"
                                        },
                                        "area_codigo": {
                                            "type": "string",
                                            "example": "DG-001"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (ej. código duplicado o resguardante inválido)"
                    },
                    "404": {
                        "description": "Área no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Areas"
                ],
                "summary": "Eliminar un área",
                "description": "Elimina un área solo si no tiene departamentos ni bienes asociados.",
                "operationId": "d8261e1575a2a594723f0f7b6e971997",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del área a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Área eliminada correctamente (Sin contenido)"
                    },
                    "409": {
                        "description": "Conflicto: No se puede eliminar porque tiene dependencias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se puede eliminar el área. Aún tiene departamentos asociados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Área no encontrada"
                    }
                }
            }
        },
        "/areas/{id}/structure": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Obtener estructura jerárquica del área",
                "description": "Retorna los departamentos del área y sus respectivas oficinas.",
                "operationId": "f07ec1866e04eb756f409e850a802b0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estructura completa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "dep_nombre": {
                                                "type": "string",
                                                "example": "Recursos Humanos"
                                            },
                                            "dep_codigo": {
                                                "type": "string",
                                                "example": "RH-01"
                                            },
                                            "oficinas": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "nombre": {
                                                            "type": "string",
                                                            "example": "Oficina de Nóminas"
                                                        },
                                                        "ofi_codigo": {
                                                            "type": "string",
                                                            "example": "NOM-01"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Área no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Area": {
                "title": "Area",
                "description": "Modelo de Área",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "area_nombre": {
                        "type": "string",
                        "example": "Recursos Humanos"
                    },
                    "area_codigo": {
                        "type": "string",
                        "example": "RH-001"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}